refactor(DEV-10393): move `<ExtendThemeProvider/>` to separate file
