+++ packages/sdk-react/src/components/receivables/InvoiceStatusChip/InvoiceStatusChip.stories.tsx
@@ -1,4 +1,4 @@
-import { ExtendThemeProvider } from '@/utils/storybook-utils';
+import { ExtendThemeProvider } from '@/utils/ExtendThemeProvider';
 import { ReceivablesStatusEnum } from '@monite/sdk-api';
 import { Alert } from '@mui/material';
 import type { Meta, StoryObj } from '@storybook/react';
