+++ packages/sdk-react/src/ui/table/TablePagination.stories.tsx
@@ -1,4 +1,4 @@
-import { ExtendThemeProvider } from '@/utils/storybook-utils';
+import { ExtendThemeProvider } from '@/utils/ExtendThemeProvider';
 import { Alert } from '@mui/material';
 import type { Meta, StoryObj } from '@storybook/react';
